+++ typescript/infra/config/environments/testnet4/gas-oracle.ts
@@ -29,8 +29,8 @@ const gasPrices: ChainMap<BigNumber> = {
   lineagoerli: ethers.utils.parseUnits('1', 'gwei'),
   polygonzkevmtestnet: ethers.utils.parseUnits('1', 'gwei'),
   chiado: ethers.utils.parseUnits('2', 'gwei'),
-  // solanadevnet: ethers.BigNumber.from('28'),
-  injective: ethers.utils.parseUnits('0.1', 'gwei'),
+  solanatestnet: ethers.BigNumber.from('28'),
+  eclipsetestnet: ethers.BigNumber.from('28'),
 };
 
 // Used to categorize rarity of testnet tokens & approximate exchange rates.
@@ -64,8 +64,8 @@ const chainTokenRarity: ChainMap<Rarity> = {
   lineagoerli: Rarity.Rare,
   polygonzkevmtestnet: Rarity.Common,
   chiado: Rarity.Common,
-  // solanadevnet: Rarity.Common,
-  injective: Rarity.Common,
+  solanatestnet: Rarity.Common,
+  eclipsetestnet: Rarity.Common,
 };
 
 // Gets the "value" of a testnet chain