+++ typescript/infra/scripts/deploy.ts
@@ -50,19 +50,13 @@ async function main() {
 
   let multiProvider = await envConfig.getMultiProvider();
 
-  // TODO: make this more generic
-  const deployerAddress =
-    environment === 'testnet4'
-      ? '0xfaD1C94469700833717Fa8a3017278BC1cA8031C'
-      : '0xa7ECcdb9Be08178f896c26b7BbD8C3D4E844d9Ba';
-
   if (fork) {
     multiProvider = multiProvider.extendChainMetadata({
       [fork]: { blocks: { confirmations: 0 } },
     });
     await useLocalProvider(multiProvider, fork);
 
-    const signer = await impersonateAccount(deployerAddress);
+    const signer = await impersonateAccount(envConfig.owners[fork].owner);
     multiProvider.setSharedSigner(signer);
   }
 
@@ -79,7 +73,6 @@ async function main() {
     );
     deployer = new HyperlaneCoreDeployer(multiProvider, ismFactory);
   } else if (module === Modules.WARP) {
-    const owner = deployerAddress;
     const injectiveRouter =
       'CCECE22C7F562110EAD32E98EA8E9B138152C49D94CDCA0554CD1959B3EEB00C';
     const ismFactory = HyperlaneIsmFactory.fromAddressesMap(
@@ -94,7 +87,7 @@ async function main() {
       deployEnvToSdkEnv[environment],
       multiProvider,
     );
-    const routerConfig = core.getRouterConfig(owner);
+    const routerConfig = core.getRouterConfig(envConfig.owners);
     config = {
       inevm: {
         ...routerConfig['inevm'],
@@ -149,7 +142,7 @@ async function main() {
     deployer = new TestQuerySenderDeployer(multiProvider);
   } else if (module === Modules.HELLO_WORLD) {
     const core = HyperlaneCore.fromEnvironment(env, multiProvider);
-    config = core.getRouterConfig(deployerAddress);
+    config = core.getRouterConfig(envConfig.owners);
     deployer = new HelloWorldDeployer(multiProvider);
   } else {
     console.log(`Skipping ${module}, deployer unimplemented`);
@@ -175,7 +168,7 @@ async function main() {
     addresses,
     verification,
     read: environment !== 'test',
-    write: true,
+    write: !fork,
   };
   // Don't write agent config in fork tests
   const agentConfig =
@@ -187,9 +180,8 @@ async function main() {
         }
       : undefined;
 
-  // prompt for confirmation
-  if ((environment === 'mainnet3' || environment === 'testnet4') && !fork) {
-    console.log(JSON.stringify(config, null, 2));
+  // prompt for confirmation in production environments
+  if (environment !== 'test' && !fork) {
     const { value: confirmed } = await prompt({
       type: 'confirm',
       name: 'value',