+++ typescript/sdk/src/consts/chains.ts
@@ -37,6 +37,8 @@ export enum Chains {
   sepolia = 'sepolia',
   solana = 'solana',
   solanadevnet = 'solanadevnet',
+  solanatestnet = 'solanatestnet',
+  eclipsetestnet = 'eclipsetestnet',
   test1 = 'test1',
   test2 = 'test2',
   test3 = 'test3',
@@ -91,6 +93,8 @@ export const Testnets: Array<CoreChainName> = [
   Chains.scrollsepolia,
   Chains.sepolia,
   Chains.solanadevnet,
+  Chains.solanatestnet,
+  Chains.eclipsetestnet,
 ];
 
 export const TestChains: Array<CoreChainName> = [